Create index.html and styles.css
These are the base files for us to work off of. index.html is our base webpage, styles.css is our base stylesheet. There are divs already set up for each section in index and classes already set up for each of those divs in styles. Change any code corresponding to your portion of the page as you like, but it's probably best to check in with the person before editing another's code. It may also be helpful for us to create our own branch to work off while we get our sections working, but that may be unnecessary.